+++ docs/changelog.md
@@ -3,6 +3,11 @@
 
 - Korean, Russian and Turkish translation added, Japanese translation fixed. Thanks to the translation team!
 - Auto build update.
+- Full support for select.lua which is a new simple command palette script embedded into mpv/libmpv.
+  In the context menu select.lua features can be found under 'View > On Screen Menu'.
+  https://github.com/mpv-player/mpv/blob/master/player/lua/select.lua
+- The helper script 'Tools\update-mpv-and-libmpv.ps1' no longer uses command line arguments,
+  it uses now the Path environment variable to find mpv and mpv.net.
 
 # v7.1.1.0 (2024-02-03)